requirements.txt
behave==1.2.6
colorama==0.4.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.3
pyyaml==6.0.1